pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soccergoals"
version = "0.1.0"
description = "Detect soccer goals in live matches and deliver video clips via Telegram"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "asyncpraw>=7.7",
    "aiosqlite<=0.17",
    "python-telegram-bot>=21",
    "yt-dlp>=2024.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["src"]
